Modularity in the program, iteration 1
- Extended the Account class to provide the functionality
for the different options of the main menu.
- Added a Bank object that encapsulates the idea of
a bank entity. The bank is currently responsible for
handling the choice from the main menu and logging in
a provided account.
- Added a MainMenu object that will handle the rendering
of the main menu, as well as getting the user's choice for it.
- Deleted the BankStatement object as unnecessary
(its functionality now resides entirely within the Account class).
- Modified main.cpp to use the new classes